Create libfeline.
Put nonstandard itostr functions in it (now templated).
Convert a ton of .c files to .cpp ones.
Change c-style casts to c++-style ones.
Create a one-variable switch in config.sh for using gcc or clang.

+++ system/libc/string/strcat.cpp
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) 2021 James McNaughton Felder
-#include <string.h>
+#include <cstring>
 
 char *strcat(char *dest, const char *src){
 	return strcpy(dest+strlen(dest), src);
@@ -32,7 +32,7 @@ size_t strlcat(char *dest, const char *src, size_t n){
 		//If we copy a null byte
 		if((*(dest_ptr+count)=*(src+count))=='\0'){
 			//We are done
-			return (dest_ptr+count)-dest;
+			return static_cast<size_t>((dest_ptr+count)-dest);
 		}
 	}
 	//Don't append a null (avoid buffer overflow)
